Sample airing data

+++ app/utils/Airing.js
@@ -402,8 +402,25 @@ export default class Airing {
     const { _id, path } = this;
     return new Promise((resolve, reject) => {
       try {
-        global.Api.delete(path);
+        if (process.env.NODE_ENV === 'production') {
+          global.Api.delete(path);
+        }
         global.RecDb.asyncRemove({ _id });
+        if (this.show && this.show.showCounts) {
+          if (this.show.showCounts.airing_count === 1) {
+            global.ShowDb.asyncRemove({ object_id: this.show.id });
+          } else {
+            global.ShowDb.asyncUpdate(
+              { object_id: this.show.id },
+              {
+                $set: {
+                  'show_counts.airing_count':
+                    this.show.showCounts.airing_count - 1
+                }
+              }
+            );
+          }
+        }
         setTimeout(() => resolve(true), 300 + 300 * Math.random());
       } catch (e) {
         console.log('Airing.delete', e);
